requirements.txt
Flask==3.0.3
langchain==0.3.7
langchain-core==0.3.15
langchain-google-genai==1.0.10
google-generativeai==0.8.3
python-dotenv==1.0.1
SpeechRecognition==3.10.1
gTTS==2.5.3
playsound==1.3.0
pyaudio==0.2.14
